// cloudreve-official-frontend/src/api/explorer.ts
export const FilePermission = {
  read: 0,
  update: 1,
  create: 2,
  delete: 3,
  share: 4,
};
// ...
export interface PermissionSettingReq {
  same_group?: string;
  everyone?: string;
  other?: string;
  anonymous?: string;
  group_explicit?: {
    [key: string]: string;
  };
  user_explicit?: {
    [key: string]: string;
  };
}